exclude non-hunt mobs from indexing
there is a bug causing name conflicts for mobs like garangatch, for which there is an arr non-hunt mob of the same name. since we filter out duplicate names, this can result in the wrong mobid being used for these mobs (which is the case for garangatch). by filtering out non-hunt mobs, we avoid that.